refactor tests
